images

+++ src/main/kotlin/ru/hse/legaltech/backend/repository/CompanyRepository.kt
@@ -3,4 +3,4 @@ package ru.hse.legaltech.backend.repository
 import org.springframework.data.jpa.repository.JpaRepository
 import ru.hse.legaltech.backend.entity.Company
 
-interface CompanyRepository : JpaRepository<Company, Long>
+interface CompanyRepository : JpaRepository<Company, Int>
